requirements.txt
Flask
numpy
Pillow
scikit-image
tensorflow
cloudinary
python-dotenv
pymongo
bcrypt
Flask-Mail 
itsdangerous
pymongo 
passlib